Add files via upload
This code creates a  NI USB-6001 DAQ session in order to track voltage for a specified period of time. It then filters the data, institutes a threshold value, and determines the number of HTR events that were detected during the session. It posts in the command window the time points of these HTR events.